Fix a dead comparator baseline, colliding conveyor keys and vein drift
The sheetmetal tank never called updateComparatorValuesPart1, so the
levels it compares against stayed at their zero initialisers for the life
of the tile. Every 144mB transfer at any non-zero level therefore re-fired
notifyNeighborsOfStateChange for the master and all nine positions across
four layers whether or not a level had changed, while a level genuinely
falling to zero notified nobody, because zero matched the stale baseline.
Part1 is now taken immediately before the drain, which is how the silo --
which has the same pair of methods -- has always used them. Part2 does not
update the baseline itself, so it has to be retaken each time.

While in that method: an empty tank was scanning all six neighbours for
redstone every tick before discovering it had nothing to push. The fluid
check now comes first.

The conveyor splitter identified its per-item routing state with
BlockPos.hashCode, which is a lossy mix of x, y and z and collides readily
across a world. Two splitters sharing a hash also share the key, so an
item still carrying the tag from the first -- it is only cleared once the
item nears its redirect target, so anything diverted or picked up keeps it
-- arrives at the second, is taken for already-routed, and is sent down
the first splitter's branch instead. Now keyed on BlockPos.toLong, which
packs the coordinates losslessly and costs the same. Items in flight
carrying an old key are simply re-assigned once.

Mineral vein selection was not reproducible. mineralList is a
LinkedHashMap and iterates in a fixed order, but it was copied into a
HashSet of entries, and MineralMix overrides neither hashCode nor equals,
so the copy ordered by identity hash -- which differs between JVM runs.
Since a vein is picked by walking that set subtracting a seeded weight,
the same uncached chunk could yield a different vein after a restart even
though the seed had not changed. A LinkedHashSet preserves the order.

The bucket wheel wrote its rotation as a float and read it back with
getInteger, which floors. The once-a-second resync was snapping the wheel
to whole degrees and comparing the truncated value against the real one.

+++ src/main/java/blusunrize/immersiveengineering/common/blocks/metal/TileEntityBucketWheel.java
@@ -169,7 +169,10 @@ public class TileEntityBucketWheel extends TileEntityMultiblockPart<TileEntityBu
 				this.digStacks.set(message.getInteger("empty"), ItemStack.EMPTY);
 			if(message.hasKey("rotation"))
 			{
-				int packetRotation = message.getInteger("rotation");
+				//Read as a float, matching the writer. getInteger on a float tag floors it, so the
+				//once-a-second resync was snapping the wheel to whole degrees and comparing a truncated
+				//value against the real one.
+				float packetRotation = message.getFloat("rotation");
 				if(Math.abs(packetRotation-rotation) > 5*IEConfig.Machines.excavator_speed)
 					rotation = packetRotation;
 			}
